Detach ActiveSession rule from AbstractDBRule.
It was the only thing using the abstraction, so I collapsed the two.
